Move modules from single module packages
