+++ src/js/common/models/collocations.ts
@@ -16,9 +16,9 @@
  * limitations under the License.
  */
 import * as Immutable from 'immutable';
-import { SrchContextType, DataRow, DataHeading } from '../../api/abstract/collocations';
-import { BacklinkWithArgs } from '../../tile';
-import { CollocMetric } from '../../../tiles/core/collocations/common';
+import { SrchContextType, DataRow, DataHeading } from '../api/abstract/collocations';
+import { BacklinkWithArgs } from '../tile';
+import { CollocMetric } from '../../tiles/core/collocations/common';
 
 
 export function ctxToRange(ctxType:SrchContextType, range:number):[number, number] {
